Port away from KTemporaryFile to QTemporaryFile
The temporary file, which is removed when the preview dialog is closed,
does not need to be created in the same place where the HTML is stored,
which (maybe) could be problematic for a direct editing of a remote URL.
So just use the standard temporary directory. With the change,
remove few additional dependencies on KUrl.