.desktop: kimagemapeditor can take a list of files
Remote URLs do not work (at least http), which could be a bug or not
(see the old bug 144151 for ftp).
For now fix the desktop file, because it specifies a MimeType entry
and that means the need for a keyword (%f, %F, %u or %U) in the Exec
line (lintian says desktop-mime-but-no-exec-code), or the MimeType
entry to be dropped.

BUG: 386696